[JAX] Check if an array is deleted when resharding it with a different device order
`jax.device_put()` that changes only the device order did not have a check on
whether the input array has been deleted or donated. When this is the case, it
would generate an error "TypeError: 'NoneType' object is not iterable" when
attempting to access `array._arrays`". Calling an explicit check makes the
error message more understandable.

PiperOrigin-RevId: 650800241